Fix crash when trying to close a closed process

+++ src/Process.cpp
@@ -98,14 +98,12 @@ NAN_METHOD(Process::TerminateProcessById) {
     int id = info[0]->NumberValue();
     HANDLE handle = OpenProcess(PROCESS_TERMINATE, false, id);
     v8::Local <v8::Function> cons = Nan::New(constructor);
-/*    try {*/
+
+    if (handle != NULL) {
         TerminateProcess(handle, 0);
         CloseHandle(handle);
-        info.GetReturnValue().Set(Nan::Undefined());
-/*    }
-    catch (const std::invalid_argument &e) {
-        info.GetReturnValue().Set(Nan::New(false));
-    }*/
+    }
+    info.GetReturnValue().Set(Nan::Undefined());
 };
 
 NAN_METHOD(Process::GetCurrentAsync) {
@@ -221,12 +219,14 @@ NAN_METHOD(Process::setToForeground) {
 };
 
 NAN_METHOD(Process::terminate) {
-        Process *obj = Nan::ObjectWrap::Unwrap<Process>(info.Holder());
-        HANDLE handle = obj->handle();
-    /*    try {*/
+    Process *obj = Nan::ObjectWrap::Unwrap<Process>(info.Holder());
+    HANDLE handle = obj->handle();
+    
+    if (handle != NULL) {
         TerminateProcess(handle, 0);
         CloseHandle(handle);
-        info.GetReturnValue().Set(Nan::Undefined());
+    }
+    info.GetReturnValue().Set(Nan::Undefined());
 };
 
 NAN_MODULE_INIT(Init) {
